+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Feature: Support S3-compatible object storage (e.g. MinIO)
 * Feature: Enable/Disable ID/Password Authentication
 * Improvement: Login Mechanism with HTTP Basic Authentication header
+* Improvement: Reactify Table Of Contents
 * Fix: Profile images are broken in User Management
 * Fix: Template page under root page doesn't work
 * Support: Upgrade libs